# swagger.json
{
    "swagger": "2.0",
    "basePath": "/",
    "info": {
        "title": "Concesionario API",
        "version": "1.0.0"
    },
    "definitions": {
        "concesionarios": {
            "type": "object",
            "properties": {
                "nombre": {
                    "type": "string"
                },
                "direccion": {
                    "type": "string"
                },

                "Coche": {
                    "type": "object",
                    "properties": {
                        "marca": {
                            "type": "string"
                        },
                        "modelo": {
                            "type": "string"
                        },
                        "cv": {
                            "type": "integer"
                        },
                        "precio": {
                            "type": "integer"
                        }
                    },
                    "required": ["marca", "modelo", "cv", "precio"]
                }
            }
        },
        "Coche": {
            "type": "object",
            "properties": {
                "marca": {
                    "type": "string"
                },
                "modelo": {
                    "type": "string"
                },
                "cv": {
                    "type": "integer"
                },
                "precio": {
                    "type": "integer"
                }
            },
            "required": ["marca", "modelo", "cv", "precio"]
        }
    },
    "paths": {
        "/concesionarios": {
            "get": {
                "summary": "Obtener todos los concesionarios",
                "operationId": "getconcesionario",
                "responses": {
                    "200": {
                        "description": "List of concesionarios",
                        "schema": {
                            "type": "array",
                            "items": {}
                        }
                    }
                }
            },
            "post": {
                "summary": "Crear un concesionario",
                "operationId": "createConcesionario",
                "consume": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "concesionarios",
                        "schema": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{message:'ok'}"
                    }
                }
            }
        },
        "/concesionarios/{id}": {
            "get": {
                "summary": "Obtener un concesionario por ID",
                "operationId": "getConcesionario",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concesionario obtenido correctamente",
                        "schema": {
                            "$ref": "#/definitions/concesionarios"
                        }
                    },
                    "404": {
                        "description": "Concesionario no encontrado"
                    }
                }
            },
            "put": {
                "summary": "Actualizar un concesionario por ID",
                "operationId": "updateConcesionario",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "concesionario",
                        "schema": {
                            "$ref": "#/definitions/concesionarios"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concesionario actualizado correctamente"
                    },
                    "404": {
                        "description": "Concesionario no encontrado"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar un concesionario por ID",
                "operationId": "deleteConcesionario",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concesionario eliminado correctamente"
                    },
                    "404": {
                        "description": "Concesionario no encontrado"
                    }
                }
            }
        },

        "/concesionarios/{id}/coches": {
            "get": {
                "summary": "Obtener todos los coches de un concesionario",
                "operationId": "getcoches",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "description": "ID del concesionario",
                "responses": {
                    "200": {
                        "description": "Coches obtenidos correctamente"
                    }
                }
            },

            "post": {
                "summary": "Agregar un nuevo coche a un concesionario",
                "operationId": "addCarToConcesionario",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "ID del concesionario"
                    }
                ],
                "requestBody": {
                    "description": "Datos del nuevo coche",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Coche agregado correctamente"
                    },
                    "400": {
                        "description": "Solicitud incorrecta, verifique el formato de los datos"
                    }
                }
            }
        },

        "/concesionarios/{id}/coches/{cocheId}": {
            "get": {
                "summary": "Obtener informacion de un coche especifico",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "ID del concesionario"
                        }
                    },
                    {
                        "in": "path",
                        "name": "cocheId",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "Posicion del coche en el array del concesionario"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Coche obtenido correctamente"
                    }
                }
            },
            "put": {
                "summary": "Actualizar informacion de un coche especifico",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "ID del concesionario"
                    },
                    {
                        "in": "path",
                        "name": "cocheId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Posicion del coche en el array del concesionario"
                    }
                ],
                "requestBody": {
                    "description": "Datos del coche a actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "marca": {
                                        "type": "string"
                                    },
                                    "modelo": {
                                        "type": "string"
                                    },
                                    "cv": {
                                        "type": "integer"
                                    },
                                    "precio": {
                                        "type": "integer"
                                    }
                                },
                                "required": ["marca", "modelo"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Coche actualizado correctamente"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar un coche especifico",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "ID del concesionario"
                    },
                    {
                        "in": "path",
                        "name": "cocheId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Posicion del coche en el array del concesionario"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Coche eliminado correctamente"
                    }
                }
            }
        }
    }
}